% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{as.MS2spectrum}
\alias{as.MS2spectrum}
\title{Convert spectra from \pkg{MSnbase} classes}
\usage{
as.MS2spectrum(x)
}
\arguments{
\item{x}{An object of class \code{\link[MSnbase:Spectrum-class]{Spectrum}}
or \code{\link[MSnbase:Spectrum-class]{Spectrum2}}}
}
\value{
An object of class \code{\linkS4class{MS2spectrum}}
}
\description{
Convert spectra from \pkg{MSnbase} classes
}
\examples{
#Load a "Spectrum2" object from MSnbase
library(MSnbase)
sp <- itraqdata[["X1"]]
#Convert this object to "MS2spectrum" class
new_sp <- as.MS2spectrum(sp)
#Or alternatively:
new_sp <- as(sp, "MS2spectrum")

}
