% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancematrix_mds.R
\name{CluMSID_MDS}
\alias{CluMSID_MDS}
\title{Multidimensional scaling of spectral similarity data}
\usage{
CluMSID_MDS(distmat, interactive = FALSE, highlight_annotated = FALSE)
}
\arguments{
\item{distmat}{A distance matrix as generated by
\code{\link{distanceMatrix}}.}

\item{interactive}{Logical, defaults to \code{FALSE}. If \code{TRUE}, an
interactive visualisation is generated using \pkg{plotly}.}

\item{highlight_annotated}{Logical, defaults to \code{FALSE}. If \code{TRUE},
points for features for which an annotation was added before using
\code{\link{distanceMatrix}} are highlighted by red colour, while other
points are grey in the MDS plot.}
}
\value{
An MDS plot generated with the help of \code{\link[stats]{cmdscale}},
  \code{\link[ggplot2]{ggplot}} and, if interactive,
  \code{\link[plotly]{ggplotly}}.
}
\description{
\code{CluMSID_MDS()} is used to generate multidimensional scaling plots from
spectral similarity data. An interactive visualisation can be produced using
\pkg{plotly}.
}
