% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{CluMSID_OPTICStbl}
\alias{CluMSID_OPTICStbl}
\title{Density-based clustering of spectral similarity data}
\usage{
CluMSID_OPTICStbl(distmat, eps = 10000, minPts = 3, eps_cl = 0.5)
}
\arguments{
\item{distmat}{A distance matrix as generated by
\code{\link{distanceMatrix}}.}

\item{eps, minPts}{OPTICS parameters, see \code{\link[dbscan]{optics}}.}

\item{eps_cl}{The reachability distance used for cluster determination,
see \code{\link[dbscan:optics]{extractDBSCAN}}.}
}
\value{
A \code{data.frame} with feature name, cluster ID and
OPTICS order for each feature in \code{distmat}.
}
\description{
\code{CluMSID_OPTICStbl()} performs density-based clustering
of spectral similarity data using the OPTICS algorithm.
}
\details{
The function internally uses \code{\link[dbscan]{optics}}
and \code{\link[dbscan:optics]{extractDBSCAN}} from the \pkg{dbscan}
package.
}
\seealso{
CluMSID_OPTICSplot
}
