% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{convertSpectrum}
\alias{convertSpectrum}
\title{Convert spectra from \pkg{MSnbase} classes}
\usage{
convertSpectrum(x)
}
\arguments{
\item{x}{An object of class \code{\link[MSnbase:Spectrum-class]{Spectrum}} or
\code{\link[MSnbase:Spectrum-class]{Spectrum2}}}
}
\value{
An object of class \code{\linkS4class{MS2spectrum}}
}
\description{
Convert spectra from \pkg{MSnbase} classes
}
