% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_merge.R
\name{mergeSpecList}
\alias{mergeSpecList}
\title{Merge list of spectra}
\usage{
mergeSpecList(speclist, tolerance = 1e-05)
}
\arguments{
\item{speclist}{A \code{list} of \code{MS2spectrum} objects to be merged.}

\item{tolerance}{The m/z tolerance to be used for merging.}
}
\value{
A \code{list} of the same length as \code{speclist} containing merged
  spectra as \code{MS2spectrum} objects. If multiple spectra contribute to
  one consensus spectrum, than this consensus spectrum is contained in the
  list multiple times at the respective positions of the contributing
  spectra.
}
\description{
\code{mergeSpecList()} is an accessory function used only inside
\code{mergeMS2spectra}.
}
\keyword{internal}
