% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_merge.R
\name{neutrallossPatterns}
\alias{neutrallossPatterns}
\title{Generate neutral loss patterns from MS2 spectra}
\usage{
neutrallossPatterns(x)
}
\arguments{
\item{x}{an object of class \code{\linkS4class{MS2spectrum}} that contains
an MS2 spectrum in the \code{spectrum} slot}
}
\value{
an object of class \code{\linkS4class{MS2spectrum}} with a neutral
    loss pattern in the \code{neutral_losses} slot
}
\description{
\code{neutrallossPatterns} generates neutral loss patterns from MS2 spectra
and adds them to \code{\linkS4class{MS2spectrum}} objects in the slot
\code{neutral_losses}.
}
\keyword{internal}
