% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\docType{class}
\name{pseudospectrum-class}
\alias{pseudospectrum-class}
\title{A custom S4 class for MS1 pseudospectra and respective
metainformation}
\description{
A custom S4 class for MS1 pseudospectra and respective
metainformation
}
\section{Slots}{

\describe{
\item{\code{id}}{a the \code{"pcgroup"} number assigned by \pkg{CAMERA}}

\item{\code{annotation}}{a character string containing a user-defined
annotation, defaults to empty}

\item{\code{rt}}{(median) retention time of the ions contained in the
pseudospectrum}

\item{\code{spectrum}}{the actual MS1 pseudospectrum as two-column
matrix (column 1 is (median) \emph{m/z}, column 2 is (median)
intensity of the ions)}
}}

