% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{accessors}
\alias{accessors}
\alias{access_id}
\alias{access_annotation}
\alias{access_precursor}
\alias{access_rt}
\alias{access_polarity}
\alias{access_spectrum}
\alias{access_neutral_losses}
\title{Accessor functions for individual slots of
\code{\linkS4class{MS2spectrum}} and
\code{\linkS4class{pseudospectrum}} objects}
\usage{
access_id(x)

access_annotation(x)

access_precursor(x)

access_rt(x)

access_polarity(x)

access_spectrum(x)

access_neutral_losses(x)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{MS2spectrum}}
or \code{\linkS4class{pseudospectrum}}}
}
\value{
The value of the respective slot of the object
    (\code{id}, \code{annotation}, \code{precursor},
    \code{rt}, \code{spectrum}, \code{neutral_losses})
}
\description{
Accessor functions for individual slots of
\code{\linkS4class{MS2spectrum}} and
\code{\linkS4class{pseudospectrum}} objects
}
\examples{
load(file = system.file("extdata",
    "annotatedSpeclist.RData",
    package = "CluMSIDdata"))

access_id(annotatedSpeclist[[1]])

load(file = system.file("extdata",
    "annotatedSpeclist.RData",
    package = "CluMSIDdata"))

access_annotation(annotatedSpeclist[[1]])

load(file = system.file("extdata",
    "annotatedSpeclist.RData",
    package = "CluMSIDdata"))

access_precursor(annotatedSpeclist[[1]])

load(file = system.file("extdata",
    "annotatedSpeclist.RData",
    package = "CluMSIDdata"))

access_rt(annotatedSpeclist[[1]])

load(file = system.file("extdata",
    "annotatedSpeclist.RData",
    package = "CluMSIDdata"))

access_polarity(annotatedSpeclist[[1]])

load(file = system.file("extdata",
    "annotatedSpeclist.RData",
    package = "CluMSIDdata"))

access_spectrum(annotatedSpeclist[[1]])

load(file = system.file("extdata",
    "annotatedSpeclist.RData",
    package = "CluMSIDdata"))

access_neutral_losses(annotatedSpeclist[[1]])

}
