% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\docType{class}
\name{MS2spectrum-class}
\alias{MS2spectrum-class}
\alias{show,MS2spectrum-method}
\title{A custom S4 class for MS2 spectra, neutral loss patterns and
respective metainformation}
\usage{
\S4method{show}{MS2spectrum}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{MS2spectrum}}}
}
\value{
Prints information from the object slots with exception of
  'spectrum' and 'neutral_losses' where only a summary is given.
}
\description{
A custom S4 class for MS2 spectra, neutral loss patterns and
respective metainformation
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: A show generic for \code{MS2spectra}.
}}

\section{Slots}{

\describe{
\item{\code{id}}{a character string similar to the ID used by XCMSonline
or the ID given in a predefined peak list}

\item{\code{annotation}}{a character string containing a user-defined
annotation, defaults to empty}

\item{\code{precursor}}{(median) \emph{m/z} of the spectrum's precursor ion}

\item{\code{rt}}{(median) retention time of the spectrum's precursor ion}

\item{\code{spectrum}}{the actual MS2 spectrum as two-column matrix
(column 1 is (median) \emph{m/z}, column 2 is (median) intensity of the
product ions)}

\item{\code{neutral_losses}}{a neutral loss pattern generated by subtracting the
product ion mass-to-charge ratios from the precursor \emph{m/z} in a
matrix format analogous to the \code{spectrum} slot}
}}

